refactor(limitAsync): move from array to promise category
limitAsync wraps an async function to limit concurrent executions, which
makes it a Promise utility rather than an array utility.

- Move src/array/limitAsync.ts (+ spec) to src/promise
- Keep a deprecated alias exported from es-toolkit/array
- Re-export the canonical symbol explicitly from the root entry so the
  name is not dropped by ambiguous star exports
- Move docs (en/ko/ja/zh_hans) under Promise Utilities
