+++ src/promise/index.ts
@@ -1,5 +1,6 @@
 export { allKeyed } from './allKeyed.ts';
 export { delay } from './delay.ts';
+export { limitAsync } from './limitAsync.ts';
 export { Mutex } from './mutex.ts';
 export { Semaphore } from './semaphore.ts';
 export { timeout } from './timeout.ts';
